moved existing components into new component file structure
