+++ example/project-nextjs-tailwind/.storybook/main.js
@@ -1,0 +1,31 @@
+const path = require('path');
+
+module.exports = {
+  stories: [
+    "../stories/**/*.stories.mdx",
+    "../stories/**/*.stories.@(js|jsx|ts|tsx)",
+  ],
+  addons: [
+    "@storybook/addon-links",
+    "@storybook/addon-essentials",
+    "@storybook/addon-interactions",
+  ],
+  framework: "@storybook/react",
+  core: {
+    builder: "@storybook/builder-webpack5",
+  },
+  webpackFinal: async (config) => {
+    config.module.rules.push({
+      test: /\.css$/i,
+      use: [
+        {
+          loader: "postcss-loader",
+          options: { implementation: require.resolve("postcss") },
+        },
+      ],
+      include: path.resolve(__dirname, "../"),
+    });
+    // Return the altered config
+    return config;
+  },
+};
