+++ example/project-nextjs-tailwind/.storybook/preview.js
@@ -1,0 +1,11 @@
+import "../styles/globals.css";
+
+export const parameters = {
+  actions: { argTypesRegex: "^on[A-Z].*" },
+  controls: {
+    matchers: {
+      color: /(background|color)$/i,
+      date: /Date$/,
+    },
+  },
+}
