+++ example/project-nextjs-tailwind/components/tokens/typography.js
@@ -1,0 +1,26 @@
+const Typography = () => {
+  return (
+    <div className="flex flex-col w-3/4 pb-3 mb-5 border-b-2">
+      <h4 className="text-center h4 h4-400 text-type-black-secondary">Typography</h4>
+      <h1 className="h1 h1-400">H1 400</h1>
+      <h1 className="h1 h1-700">H1 700</h1>
+      <h2 className="h2 h2-400">H2 400</h2>
+      <h2 className="h2 h2-700">H2 700</h2>
+      <h3 className="h3 h3-400">H3 400</h3>
+      <h3 className="h3 h3-700">H3 700</h3>
+      <h4 className="h4 h4-400">H4 400</h4>
+      <h4 className="h4 h4-700">H4 700</h4>
+      <h5 className="h5 h5-400">H5 400</h5>
+      <h5 className="h5 h5-700">H5 700</h5>
+      <h5 className="h6 h5-400">H5 400</h5>
+      <h6 className="h6 h6-700">H6 700</h6>
+      <p className="400">400</p>
+      <p className="500">500</p>
+      <p className="600">600</p>
+      <p className="700">700</p>
+
+    </div>
+  );
+};
+
+export default Typography;
